% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Available_Areas_Noise_Analysis_Rooftops.R
\name{availableAreaHubs_noiseAnalysis}
\alias{availableAreaHubs_noiseAnalysis}
\title{availableAreaHubs_noiseAnalysis Function}
\usage{
availableAreaHubs_noiseAnalysis(p_schools_uni, p_grave_embassy,
  p_hospitals, p_park_reserve, p_flight_paths, p_ny_neigh,
  p_buildings_manhattan, p_noise_map_raster)
}
\arguments{
\item{p_schools_uni}{This is the path where the shapefile of schools and universities is located}

\item{p_grave_embassy}{This is the path where the shapefile of graveyards and embassies is located}

\item{p_hospitals}{This is the path where the shapefile of hospitals is locates}

\item{p_park_reserve}{This is the path where the shapefile of parks and reserve areas is located}

\item{p_flight_paths}{This is the path where the shapefile of the flight restricted areas (Due to the presence of airports)is located}

\item{p_ny_neigh}{This is the path where the shapefile of border of New York or Manhattan (It depends of the processing time considered the selection of the whole New York or only Manhattan) district is located}

\item{p_buildings_manhattan}{This is the path where the shapefile of the rooftops of the buildings is located (It could be the buildings of the whole New York, Manhattan or a sample. It depends on the processing time considered)}

\item{p_noise_map_raster}{This is the path where the noise map in raster format is located}
}
\value{
summary_stats_build The function returns a data frame with the noise statistics presented in each of the roofs of the selected buildings
}
\description{
This function is responsible for taking the main sources of information to
calculate the area available for the flight of air-taxis over New York City (Manhattan).
For this, it is responsible for carrying out the calculation of the restricted
areas considering the presence of Schools, Parks, Embassies, Cemeteries, Hospitals and
Flight paths. Once this area is estimated, the function is responsible for obtaining
what are the roofs of the benefits that are on this available area. Having these
potential roofs, the crossing is made with the noise map of the city, which is
very important for the establishment of the hubs.
The function returns the noise statistics of each of the roofs of the buildings.
}
\examples{
path_schools_uni <- system.file("extdata/final_schools_uni_man.shp", package = "RooftopsHubsAnalyisisLidar")
path_grave_embassy <- system.file("extdata/final_graveyard_embassy_man.shp", package = "RooftopsHubsAnalyisisLidar")
path_hospitals <- system.file("extdata/final_hospitals_man.shp", package = "RooftopsHubsAnalyisisLidar")
path_park_reserve <- system.file("extdata/final_park_reserve_man.shp", package = "RooftopsHubsAnalyisisLidar")
path_fligh_paths <- system.file("extdata/final_flightpaths_man.shp", package = "RooftopsHubsAnalyisisLidar")
path_ny_neigh <- system.file("extdata/boundaries_manhattan.shp", package = "RooftopsHubsAnalyisisLidar")
buildings_manhattan <- system.file("extdata/sample_buildings_manhattan.shp", package = "RooftopsHubsAnalyisisLidar")
noise_map_raster <- system.file("extdata/noise_m.tif", package = "RooftopsHubsAnalyisisLidar")
availableAreaHubs_noiseAnalysis(path_schools_uni, path_grave_embassy, path_hospitals, path_park_reserve, path_fligh_paths, path_ny_neigh, buildings_manhattan, noise_map_raster)
}
\keyword{area}
\keyword{available}
\keyword{buildings,}
\keyword{clip,}
\keyword{lidar,}
\keyword{plot,}
\keyword{rooftops,}
\keyword{shapefile,}
