% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_Rooftop_Shape_Analysis_Lidar.R
\name{buildings_grid_metrics_lidar}
\alias{buildings_grid_metrics_lidar}
\title{buildings_grid_metrics_lidar Function}
\usage{
buildings_grid_metrics_lidar(list_elements_lidar, size_pixel = 5)
}
\arguments{
\item{list_elements_lidar}{List with the generated lidar files to perform the metric statistics plots of each one of them}

\item{size_pixel}{Desired pixel size to estimate the average heights in each of the buildings, Default 5m}
}
\description{
Function to plot the mean of the elevations as a grid object
}
\examples{
library(RColorBrewer)
library(stringr)
path_folder_las <- system.file("extdata/sample_las", package = "RooftopsHubsAnalyisisLidar")
list_las <- base::list.files(path_folder_las,full.names = TRUE)
buildings_grid_metrics_lidar(list_las)
}
