% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Available_Areas_Noise_Analysis_Rooftops.R
\name{make_dissolve_one_f}
\alias{make_dissolve_one_f}
\title{make_dissolve_one_f Function}
\usage{
make_dissolve_one_f(layer_obj, field_diss)
}
\arguments{
\item{layer_obj}{Path of the shapefile to perform the validation}

\item{field_diss}{Name of the field which is going to be used to perform the dissolve}
}
\description{
Function to perform the dissolve of a shapefile according to a certain field
}
