% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Available_Areas_Noise_Analysis_Rooftops.R
\name{noise_statistics_per_building}
\alias{noise_statistics_per_building}
\title{noise_statistics_per_building Function}
\usage{
noise_statistics_per_building(noise_map, shape_buildings_p)
}
\arguments{
\item{noise_map}{Noise map in raster format}

\item{shape_buildings_p}{Shapefile of the potential building for develpoing Air-Taxis Hubs}
}
\description{
Function to perform the calculation of the noise pixel statistics for each building
}
