% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_Rooftop_Shape_Analysis_Lidar.R
\name{verify_proj_CRS_lidar_building}
\alias{verify_proj_CRS_lidar_building}
\title{verify_proj_CRS_lidar_building Function}
\usage{
verify_proj_CRS_lidar_building(path_lidar_file, path_shp_buildings)
}
\arguments{
\item{path_lidar_file}{Path of the lidar file (.las) to perform the validation}

\item{path_shp_buildings}{Path of the shapefile to perform the validation}
}
\description{
Function that verifies if the lidar file and the shapefile have the same coordinate system
}
